Fix yarn lint:code not work for typescript bug & Refine eslint config & Upgrade deps
