+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "react-cool-starter",
-  "version": "3.2.2",
+  "version": "3.2.3",
   "description": "A starter boilerplate for an universal web application with the best development experience and best practices.",
   "main": "index.ts",
   "engines": {
@@ -92,7 +92,7 @@
       "command": "tsc --noEmit"
     },
     "lint:code": {
-      "command": "eslint --fix ."
+      "command": "eslint --fix . --ext .js,.jsx,.ts,.tsx"
     },
     "lint:style": {
       "command": "stylelint --fix \"./**/*.css\" \"./**/*.scss\" \"./**/*.sass\" \"!coverage\""
@@ -157,6 +157,7 @@
       "airbnb/hooks",
       "plugin:@typescript-eslint/recommended",
       "plugin:prettier/recommended",
+      "prettier/@typescript-eslint",
       "prettier/react"
     ],
     "plugins": [
@@ -180,6 +181,8 @@
     },
     "rules": {
       "global-require": "off",
+      "no-console": "off",
+      "no-underscore-dangle": "off",
       "react/prop-types": "off",
       "react-hooks/exhaustive-deps": "off",
       "react/jsx-props-no-spreading": "off",
@@ -187,10 +190,10 @@
         "error",
         {
           "extensions": [
-            ".ts",
-            ".tsx",
             ".js",
-            ".jsx"
+            ".jsx",
+            ".ts",
+            ".tsx"
           ]
         }
       ],
@@ -299,7 +302,7 @@
     "@loadable/component": "^5.10.3",
     "@loadable/server": "^5.10.3",
     "asset-require-hook": "^1.2.0",
-    "autoprefixer": "^9.7.0",
+    "autoprefixer": "^9.7.1",
     "axios": "^0.19.0",
     "babel-plugin-dynamic-import-node": "^2.3.0",
     "babel-plugin-transform-remove-console": "^6.9.4",
@@ -307,7 +310,7 @@
     "chalk": "^2.4.2",
     "compression": "^1.7.4",
     "connected-react-router": "^6.5.2",
-    "core-js": "^3.3.5",
+    "core-js": "^3.3.6",
     "css-modules-require-hook": "^4.2.3",
     "eslint-plugin-react-hooks": "^1.7.0",
     "express": "^4.17.1",
@@ -338,21 +341,21 @@
     "@types/compression": "^1.0.1",
     "@types/enzyme": "^3.10.3",
     "@types/enzyme-adapter-react-16": "^1.0.5",
-    "@types/express": "^4.17.1",
-    "@types/helmet": "^0.0.44",
+    "@types/express": "^4.17.2",
+    "@types/helmet": "^0.0.45",
     "@types/hpp": "^0.2.1",
-    "@types/jest": "^24.0.20",
+    "@types/jest": "^24.0.21",
     "@types/loadable__component": "^5.10.0",
     "@types/loadable__server": "^5.9.1",
     "@types/morgan": "^1.7.37",
     "@types/react": "^16.9.11",
     "@types/react-dev-utils": "^9.0.1",
     "@types/react-dom": "^16.9.3",
-    "@types/react-helmet": "^5.0.13",
+    "@types/react-helmet": "^5.0.14",
     "@types/react-redux": "^7.1.5",
     "@types/react-router": "^5.1.2",
     "@types/react-router-config": "^5.0.1",
-    "@types/react-router-dom": "^5.1.0",
+    "@types/react-router-dom": "^5.1.2",
     "@types/react-test-renderer": "^16.9.1",
     "@types/redux-mock-store": "^1.0.1",
     "@types/serialize-javascript": "^1.5.0",
@@ -383,7 +386,7 @@
     "jest": "^24.9.0",
     "lint-staged": "^9.4.2",
     "mini-css-extract-plugin": "^0.8.0",
-    "nock": "^11.6.0",
+    "nock": "^11.7.0",
     "nodemon": "^1.19.4",
     "npm-run-all": "^4.1.5",
     "optimize-css-assets-webpack-plugin": "^5.0.3",
@@ -407,7 +410,7 @@
     "url-loader": "^2.2.0",
     "webpack": "^4.41.2",
     "webpack-bundle-analyzer": "^3.6.0",
-    "webpack-cli": "^3.3.9",
+    "webpack-cli": "^3.3.10",
     "webpack-dev-middleware": "^3.7.2",
     "webpack-hot-middleware": "^2.25.0",
     "webpack-manifest-plugin": "^2.2.0"
